--- modules/variational/examples/FCC/fcc.gfg
@@ -1,30 +1,0 @@
-0.000000	0.000000	0.000000	0.5	1
-0.000000	0.000000	1.000000	0.5	1
-0.000000	1.000000	0.000000	0.5	1
-1.000000	0.000000	0.000000	0.5	1
-1.000000	0.000000	1.000000	0.5	1
-1.000000	1.000000	0.000000	0.5	1
-1.000000	1.000000	1.000000	0.5	1
-0.000000	0.500000	0.500000	0.5	1
-0.000000	0.500000	1.500000	0.5	1
-0.000000	1.500000	0.500000	0.5	1
-0.000000	1.500000	1.500000	0.5	1
-1.000000	0.500000	0.500000	0.5	1
-1.000000	0.500000	1.500000	0.5	1
-1.000000	1.500000	0.500000	0.5	1
-1.000000	1.500000	1.500000	0.5	1
-0.500000	0.000000	0.500000	0.5	1
-0.500000	0.000000	1.500000	0.5	1
-0.500000	1.000000	0.500000	0.5	1
-0.500000	1.000000	1.500000	0.5	1
-1.500000	0.000000	0.500000	0.5	1
-1.500000	0.000000	1.500000	0.5	1
-1.500000	1.000000	0.500000	0.5	1
-1.500000	1.000000	1.500000	0.5	1
-0.500000	0.500000	0.000000	0.5	1
-0.500000	0.500000	1.000000	0.5	1
-0.500000	1.500000	0.000000	0.5	1
-0.500000	1.500000	1.000000	0.5	1
-1.500000	0.500000	0.000000	0.5	1
-1.500000	0.500000	1.000000	0.5	1
-1.500000	1.500000	0.000000	0.5	1
