--- modules/variational/examples/FCC/fcc.unq
@@ -1,12 +1,0 @@
-0.000002	0.000002	0.500000	0.307088
-0.000002	0.500000	0.000002	0.307088
-0.500000	0.500000	0.500000	0.307088
-0.500000	0.000002	0.000002	0.307088
-0.250000	0.250000	0.750000	0.170258
-0.250000	0.750000	0.750000	0.170068
-0.250000	0.250000	0.250000	0.170453
-0.750000	0.250000	0.750000	0.170068
-0.750000	0.750000	0.750000	0.169884
-0.750000	0.750000	0.250000	0.170068
-0.750000	0.250000	0.250000	0.170258
-0.250012	0.750000	0.250000	0.170258
