--- modules/variational/examples/FCC/fcc_cell.gfg
@@ -1,4 +1,0 @@
-0	0	0	0.707106781186548	1
-0	0.5	0.5	0.707106781186548	1
-0.5	0	0.5	0.707106781186548	1
-0.5	0.5	0	0.707106781186548	1
