--- modules/variational/examples/FCC/fcc_var.gfg
@@ -1,127 +1,0 @@
-0	1	1	0.01	1.0
-0.0122236	1.00286	1.00176	0.01	1.0
-0.0245385	1.00573	1.00354	0.01	1.0
-0.0368534	1.00861	1.00531	0.01	1.0
-0.0491683	1.01149	1.00709	0.01	1.0
-0.0614832	1.01436	1.00886	0.01	1.0
-0.0737979	1.01724	1.01064	0.01	1.0
-0.0861127	1.02012	1.01242	0.01	1.0
-0.0984273	1.023	1.01419	0.01	1.0
-0.110742	1.02587	1.01597	0.01	1.0
-0.123056	1.02875	1.01774	0.01	1.0
-0.135371	1.03163	1.01952	0.01	1.0
-0.147685	1.03451	1.0213	0.01	1.0
-0.16	1.03739	1.02308	0.01	1.0
-0.172314	1.04027	1.02485	0.01	1.0
-0.184628	1.04314	1.02663	0.01	1.0
-0.196942	1.04602	1.02841	0.01	1.0
-0.209255	1.0489	1.03019	0.01	1.0
-0.221569	1.05178	1.03197	0.01	1.0
-0.233882	1.05466	1.03375	0.01	1.0
-0.246195	1.05755	1.03553	0.01	1.0
-0.258507	1.06043	1.03732	0.01	1.0
-0.270819	1.06331	1.0391	0.01	1.0
-0.28313	1.0662	1.04089	0.01	1.0
-0.295439	1.06909	1.04268	0.01	1.0
-0.307748	1.07199	1.04447	0.01	1.0
-0.320055	1.07489	1.04626	0.01	1.0
-0.332359	1.0778	1.04806	0.01	1.0
-0.344661	1.08071	1.04986	0.01	1.0
-0.356959	1.08364	1.05166	0.01	1.0
-0.369253	1.08659	1.05347	0.01	1.0
-0.381542	1.08955	1.05528	0.01	1.0
-0.393824	1.09254	1.0571	0.01	1.0
-0.406098	1.09556	1.05893	0.01	1.0
-0.418362	1.09861	1.06075	0.01	1.0
-0.430614	1.10172	1.06259	0.01	1.0
-0.442852	1.10487	1.06443	0.01	1.0
-0.455072	1.10809	1.06627	0.01	1.0
-0.467272	1.11138	1.06811	0.01	1.0
-0.479447	1.11476	1.06996	0.01	1.0
-0.491594	1.11825	1.0718	0.01	1.0
-0.503708	1.12185	1.07364	0.01	1.0
-0.515783	1.12558	1.07547	0.01	1.0
-0.527813	1.12946	1.07729	0.01	1.0
-0.539793	1.1335	1.0791	0.01	1.0
-0.551714	1.13772	1.08088	0.01	1.0
-0.56357	1.14213	1.08263	0.01	1.0
-0.575352	1.14675	1.08435	0.01	1.0
-0.587052	1.15158	1.08602	0.01	1.0
-0.598663	1.15665	1.08765	0.01	1.0
-0.610175	1.16195	1.08922	0.01	1.0
-0.621582	1.16749	1.09073	0.01	1.0
-0.632877	1.17327	1.09216	0.01	1.0
-0.644054	1.1793	1.09351	0.01	1.0
-0.65511	1.18556	1.09477	0.01	1.0
-0.666045	1.19206	1.09594	0.01	1.0
-0.676859	1.19877	1.09699	0.01	1.0
-0.687558	1.20568	1.09793	0.01	1.0
-0.69815	1.21276	1.09875	0.01	1.0
-0.708647	1.22	1.09943	0.01	1.0
-0.719063	1.22737	1.09997	0.01	1.0
-0.729416	1.23484	1.10036	0.01	1.0
-0.739726	1.24237	1.10059	0.01	1.0
-0.750014	1.24994	1.10068	0.01	1.0
-0.760302	1.2575	1.1006	0.01	1.0
-0.770613	1.26504	1.10037	0.01	1.0
-0.780967	1.2725	1.09999	0.01	1.0
-0.791384	1.27987	1.09946	0.01	1.0
-0.801882	1.28711	1.09879	0.01	1.0
-0.812476	1.29419	1.09798	0.01	1.0
-0.823177	1.3011	1.09705	0.01	1.0
-0.833993	1.30781	1.096	0.01	1.0
-0.844929	1.3143	1.09484	0.01	1.0
-0.855988	1.32056	1.09358	0.01	1.0
-0.867166	1.32659	1.09224	0.01	1.0
-0.878463	1.33237	1.09081	0.01	1.0
-0.889871	1.33791	1.0893	0.01	1.0
-0.901384	1.34321	1.08773	0.01	1.0
-0.912996	1.34827	1.08611	0.01	1.0
-0.924697	1.3531	1.08443	0.01	1.0
-0.936479	1.35772	1.08272	0.01	1.0
-0.948335	1.36213	1.08097	0.01	1.0
-0.960257	1.36635	1.07919	0.01	1.0
-0.972236	1.37039	1.07738	0.01	1.0
-0.984266	1.37427	1.07556	0.01	1.0
-0.99634	1.37801	1.07373	0.01	1.0
-1.00845	1.38161	1.07189	0.01	1.0
-1.0206	1.38509	1.07004	0.01	1.0
-1.03277	1.38848	1.0682	0.01	1.0
-1.04497	1.39177	1.06635	0.01	1.0
-1.05719	1.395	1.06451	0.01	1.0
-1.06943	1.39815	1.06266	0.01	1.0
-1.08168	1.40126	1.06083	0.01	1.0
-1.09394	1.40431	1.059	0.01	1.0
-1.10621	1.40734	1.05717	0.01	1.0
-1.1185	1.41033	1.05535	0.01	1.0
-1.13078	1.4133	1.05354	0.01	1.0
-1.14308	1.41624	1.05173	0.01	1.0
-1.15537	1.41918	1.04992	0.01	1.0
-1.16767	1.4221	1.04812	0.01	1.0
-1.17998	1.42501	1.04632	0.01	1.0
-1.19228	1.42792	1.04452	0.01	1.0
-1.20459	1.43081	1.04273	0.01	1.0
-1.2169	1.43371	1.04094	0.01	1.0
-1.22921	1.4366	1.03915	0.01	1.0
-1.24152	1.43949	1.03736	0.01	1.0
-1.25383	1.44238	1.03558	0.01	1.0
-1.26614	1.44526	1.03379	0.01	1.0
-1.27845	1.44815	1.03201	0.01	1.0
-1.29077	1.45103	1.03023	0.01	1.0
-1.30308	1.45391	1.02844	0.01	1.0
-1.31539	1.4568	1.02666	0.01	1.0
-1.3277	1.45968	1.02488	0.01	1.0
-1.34002	1.46256	1.0231	0.01	1.0
-1.35233	1.46544	1.02132	0.01	1.0
-1.36464	1.46833	1.01954	0.01	1.0
-1.37696	1.47121	1.01776	0.01	1.0
-1.38927	1.47409	1.01599	0.01	1.0
-1.40158	1.47697	1.01421	0.01	1.0
-1.4139	1.47985	1.01243	0.01	1.0
-1.42621	1.48274	1.01065	0.01	1.0
-1.43852	1.48562	1.00887	0.01	1.0
-1.45084	1.4885	1.0071	0.01	1.0
-1.46315	1.49138	1.00532	0.01	1.0
-1.47546	1.49426	1.00354	0.01	1.0
-1.48778	1.49714	1.00176	0.01	1.0
-1.5	1.5	1	0.01	1.0
