+++ modules/variational/examples/random/Makefile
@@ -1,0 +1,3 @@
+# A single variational curve is mapped in a small, random configuration
+random: random.cc
+	g++ random.cc -o random -I${VACUUMMS_ROOT}/include -lm -lftw_general -lvacuumms_variational -lvacuumms_cpp -L${VACUUMMS_ROOT}/lib -Wl,-rpath,${VACUUMMS_ROOT}/lib
